修复每日推荐第一次加载的时候会把数据删除了的问题

+++ app/src/main/java/cn/xybbz/config/recommender/RecentHistoryCache.kt
@@ -26,7 +26,8 @@ class RecentHistoryCache(
             MusicDataTypeEnum.RECOMMEND,
             connectionId
         )
-        db.musicDao.removeByType(MusicDataTypeEnum.RECOMMEND, itemIds = songIds)
+        if (songIds.isNotEmpty())
+            db.musicDao.removeByType(MusicDataTypeEnum.RECOMMEND, itemIds = songIds)
     }
 
     suspend fun contains(items: List<XyMusic>): List<XyMusic> = withContext(Dispatchers.IO) {
